Use compiled regular expression for efficiency
The implementation of the EDS file loader uses regular expressions to
determine if the found index is a main or a sub index.

It is possible to to gain a significant speedup when pre-compiled
regular expressions are used instead of constantly recompiling the
regular expressions on the fly.

The used regular expressions are now pre-compiled in the `eds.py`
module.