Refactoring: Class hierarchies for PDO nodes
As with the rest of the CANopen framework, we have to distinguish two
distinct situations: is the node we are dealing with a remote or a local
node?

Is the node a remote node then we don't own any of the process data. We
merely subscribe to the PDO COBIDs and read/serve them on the network.

Is the node a local node then the canopen library is representing this
node on the network. This means the data sent and received is our
private data. Thus reading and writing the data does not involve the
network, but reads the data from the internal data store of the class
which is continuously updated upon reception of new RPDOs. Likewise data
sent via TPDOs is also taken from the internal data store.