Move the data store into node base class
The data store is the place that holds the current data of the node as
defined by the object dictionary.

In addition to this data the node also needs to know who needs to be
notified if part of the data changes. This is done via the so called
data store traps.