Only allow byte wise mapping
The CANopen standard allows the bitwise mapping of data. In the current
implementation we receive data in bytes and store them in bytes.
Handling the mapping in bit level complicated the map handling. For now
only bytewise mapping is supported. Therefore the mapping length is
divided by 8 to yield the map length in bytes, not bits.