Get rid of excess Variable, Record and Array classes
Throughout the sources multiple definitions of the classes Variable,
Record and Array were defined. All of these were very similar in nature
and shared the same ideas.

The main concept was the following: provide a look into the object
dictionary and a possibility to set and get the stored data.

All of this functionality has been moved into the objectdictionary
subpackage. Data can be read and written via convenience functions
defined in the Node class.

The API provides a total of 4 access options: raw, value, phys and bits.

    * Raw returns the raw byte data
    * Value returns the unpacked underlying Python data
    * Phys returns scaled data
    * Bits allows bit wise access