Provide an implementation for remote PDO nodes
Remote PDO nodes should be listening only, because they do not own the
data sent around on the network. Thus the receive AND transmit values
are treated as receive PDOs.

This will subscribe to the network for the corresponding COB-IDs and
allow to register callbacks, but avoids complicated object dictionary
variable guarding and updating.