Change the PDO subscription system
In previous versions of this library only one subscription per COB-ID
was supported. In case of NMT and SDO this is the correct behaviour, but
it is absolutely possible and allowed that more than one PDO listens for
a certain COB-ID. Thus in case of PDO the subscriptions need to be a
dictionary of lists.