Allow overriding the subindex in pdo.Map.add_variable().
If the keyword argument "subindex" is given and non-zero, a mapping
entry is generated for the given subindex instead of the one in the
object dictionary.

Background: Although such mappings should not be valid, for some
devices it is necessary to include them when sending PDO mappings.
The one this was tested with has a fixed number of eight entries,
corresponding to each PDO data byte.  Mapping e.g. 16-bit variables is
done by entering the object index twice, with subindex 00 and 01,
respectively.  So by forcing the subindex=01, it creates a virtual OD
entry corresponding to the high byte of the 16-bit integer object.

This is obviously not convered by the standard and only needed for
broken devices.  Tested with AC motor controllers from Curtis
Instruments, Inc.