remove unimplemented methods
After a deep dive into the code to better understand the mechanism already implemented I think the best approach to handle the PDO's communication is through the remote and local nodes since it differs a bit the behavior when handling local or remote nodes. So, in my opinion we can implement in the local node a self handler mechanism to read and set (affect) the variables in the Object dictionary. In the remote node the case is a bit different because in depends much more on the application and in what we want to do with the remote node.