solved access problem to PDOs and add new feature
The TPDOs and RPDOs were colliding  since both could have the same key when merging the maps (e.g., RPDO[1] and TPDO[1] have the same key).
So a new way of accessing the PDO was implemented and when merging both RPDOs and TPDOs each will have a unique identifier based in the CANOpen specification. So now we can access the a PDO like PDO[0x1A00] that means accessing the TPDO[1] or PDO.tx[1].