Reading DefaultValue from EDS is wrong for data type OCTET_STRING and DOMAIN.
The DefaultValue for data type OCTET_STRING and DOMAIN is formatted with two
hex digits for each byte in the EDS - see corresponding CiA standard. Also
the content of the default value has to end up in a bytes instance and not in
a string instance. (otherwise an exception is thrown when the SDOServer tries
to answer a SDO upload request for an object dictionary entry with this data type.)