ds402: Restore operation mode after homing only on explicit request.
Add a new parameter restore_op_mode which defaults to False, and skip
changing back to the previous mode unless it is explicitly enabled by
passing True.  Note that most applications will decide on the needed
mode after homing and therefore do not need this behavior, hence the
new default.