ds402: Support checking the homing status.
Implement a new method BaseNode402.is_homed() which will switch to the
corresponding operation mode HOMING and then examine the Statusword
for the homing status bits.

This allows to skip the homing procedure in case the drive controller
still knows its position, but e.g. the Python application was
restarted.