p402: Use TPDO to get the operation mode if possible.
Fall back to the previous behavior using SDO if the relevant object
0x6061 is not mapped to a TPDO.  The property getter still blocks
until an up-to-date value was received, by waiting for the respective
TPDO up to a configurable timeout of 0.2 seconds by default.  If the
TPDO does not look like it will be transmitted regularly (from its
is_periodic property), the method will not block and just return the
last received TPDO's value.

A lookup cache tpdo_pointers is added to keep track of the needed
pdo.Map instance, analog to the rpdo_pointers.