p402: Do not switch any states during setup_402_state_machine().
Reading the PDO configuration is possible in OPERATIONAL or
PRE-OPERATIONAL states, so switching that is unnecessary.  The
application should be responsible to handle such transitions, and the
library function should be usable without disturbing the application
logic.

Changing the DS402 state machine to SWITCH ON DISABLED is also not
necessary.  The drive may be in whatever state from a previous usage,
and then the change to SWITCH ON DISABLED may even trigger an
exception because there is no way to reach it directly.  So this
transition should also be the application's responsibility.