p402: Make reading the PDO configuration optional during setup.
If the application already configured the PDOs and called .save() on
the pdo.Maps object, there is no sense in reading everything back
again in the setup_pdos() method.  Provide an optional argument to
disable that behavior.

A call to subscribe to the PDOs from the network is added because that
side-effect of pdo.read() is necessary for the TPDO callback to work.