p402: Add blocking method check_statusword().
In contrast to the property getter, this method will not simply return
the cached last value, but actually wait for an updated TPDO
reception.  If no TPDO is configured, or it is not expected
periodically, the usual statusword getter takes over and returns
either the last received value or queries via SDO as a fallback.

The timeout for receiving a TPDO can be configured individually per
call, defaulting to 0.2 seconds (TIMEOUT_CHECK_TPDO).