p402: Forbid transitions to states which only the drive can enter.
The NOT READY TO SWITCH ON and FAULT REACTION ACTIVE states can only
be reached if the drive encounters an error or is still busy with its
self-testing.  Raise an exception in case these states are requested
via the property.

In extension, only the drive can trigger a transition to the FAULT
state, so that is never valid as an end goal requested by the user.