p402: Adjust automatic state transition paths.
As there is a direct transition from QUICK STOP ACTIVE to OPERATION
ENABLED, remove that from the transition paths.  Instead go through
the SWITCH ON DISABLED state, closing the cycle to make it work for
anything between SWITCH ON DISABLED and OPERATION ENABLED.

Also remove the self-reference OPERATION ENABLED to itself, which is
useless.  The whole state changing code will only be called if the
target state and the current state do not match.