DS402: Allow handling the operation mode via PDO. (#257)
* pdo: Document relation between start() and the period attribute.

* pdo: Add a property to check if period updates are transmitted.

* p402: Skip sending the controlword RPDO if configured with a period.

Transmitting the RPDO only makes sense if the value takes effect
immediately.  If there is already a cyclic task configured, or the
transmission type configuration requires a SYNC to be sent for the
change to apply, there is no sense in sending the PDO automatically in
the controlword setter.

* p402: Use RPDO to set the operation mode if possible.

Fall back to the previous behavior using SDO if the relevant object
0x6060 is not mapped to an RPDO.

* p402: Use TPDO to get the operation mode if possible.

Fall back to the previous behavior using SDO if the relevant object
0x6061 is not mapped to a TPDO.  The property getter still blocks
until an up-to-date value was received, by waiting for the respective
TPDO up to a configurable timeout of 0.2 seconds by default.  If the
TPDO does not look like it will be transmitted regularly (from its
is_periodic property), the method will not block and just return the
last received TPDO's value.

A lookup cache tpdo_pointers is added to keep track of the needed
pdo.Map instance, analog to the rpdo_pointers.

* p402: Improve documentation on PDO tracking dicts.

Consistently use empty dict literal for initialization.  Provide more
useful comments about the expected contents.

* p402: Check PDO configuration for the Operation Mode objects.

Switching operation modes for several drives simultaneously must be
done via PDO.  There is still a fallback mechanism via SDO, but a
warning should be issued when that is about to be used.

Co-authored-by: André Filipe Silva <af-silva@users.noreply.github.com>