p402: Fix missing fallback behavior in check_statusword(). (#277)
The docstring for BaseNode402.check_statusword() promises it will fall
back to SDO if the TPDO is not configured as periodic.  However, that
actually only happens when there is no TPDO for the Statusword at all.
Fix that small detail, as the code using this function relies on
getting an up-to-date value and on the implicit throttling caused by
the SDO round-trip time.