eds: Fix exception in comment parsing with hexadecimal line count. (#279)
Inside an EDS file's [Comments] section, the Lines attribute may be
expressed as a hexadecimal number prefixed with 0x.  The
ConfigParser.getint() method however does not support such a format.

Switch to the default int() constructor with a base=0 argument to
enable auto-detection of this condition.  The same is done for
basically every other numeric value in the EDS importer.