Use smaller of response size and size in OD
Some devices use simplified logic for sending SDO response.
They send an expedited 4 byte response for all sizes up to U32/I32.
Unpacking code expects data size to exactly match the size in the
object dictionary.
Therefore use the smaller of the size for the variable in the OD and
the actual response size.