Mirror some OD access API to the SdoBase / SdoVariable classes.
The ultimate goal here is being able to switch from a "live" mode to
an "offline" mode by accessing either the SDO client or the object
dictionary while accessing the same methods.

The convenient get_variable() method is mirrored from the OD, raising
no exceptions and encapsulating the two-staged access for sub-objects
in records or arrays.

The writable and readable properties on an SdoVariable are simply
passed through to the underlying OD properties.  This is solely a
convenience to match the ODVariable API without having to mention the
internal sdo_var.od attribute.