Migrate setup.cfg to pyproject.toml for newer setuptools compat.
The setuptools>=69 dependency includes a fix to automatically include
the py.typed marker in the distribution, thus making type hints usage
much easier.

The newer version however fails to build when the basic project
metadata is missing in pyproject.toml.  Thus the settings from
setup.cfg are migrated entirely, upgrading to setuptools_scm >= 8 in
the process.