Revert "change missing Baudrate to BaudRate" (#452)
The CiA 306 CANopen electronic data sheet (EDS) standard is quite clear
that in the DeviceCommisioning block, baudrate is spelled 'Baudrate' and
not 'BaudRate'. This commit is to be reverted, as it prevents loading
of standards compliant DCF files.

Chapter 5.3.3.3 is clear in the casing of Baudrate for
DeviceCommisioning and BaudRate for DeviceInfo. This seems to be an
inconcistency in the standard, and we should probably accept both as a
long term solution.

This reverts commit 074cbbc4d4c2c7803dc1e0151f6f534dbe44efc8.