Fix capitalization of Baudrate in [DeviceComissioning] section.
As pointed out in #452, there is an inconsistency in the CiA 306 spec
regarding the exact spelling.  Under DeviceComissioning (sic!), it is
spelled with a lowercase "r", while under DeviceInfo it is BaudRate_*
with an uppercase "R".

That change however broke the tests, because the library currently
handles keys in a case-sensitive way (which is wrong in itself).