Add a _DummyNetwork class to designate "uninitialized" networks.
This class can replace the dummy "None" value for attribute
initializations, which can then be properly typed as Network to avoid
static type checking errors.

Switch to that new initialization in the NmtBase class as an example.

This has the benefit of not needing `self.network is not None` checks
at run-time wherever a method or attribute access is used, but still
satisfies static type checking.  When hitting such a code path at
run-time, of course it will lead to an exception because the
attributes required in the Network methods are not set.  But that is a
case of wrong API usage (accessing a network without associating it
first), which a static checker cannot detect reliably.  The dummy
class could be modified to provide a better exception message if
desired, but this is just a minimal proof of concept so far.