Skip the "highest subindex" entry when iterating over SdoRecord.
The count entry itself is not part of the data (according to CiA 301),
thus should not be yielded from an iterator.  That matches the
behavior of SdoArray, which also yields only the array contents.

Note that the basis of returned record sub-objects is still the subset
described by the OD, which might be smaller than the actual entries
accessible on the node, and less than indicated by the record's
subindex 0.  Thus the count (and iteration set) is reduced by one
element only if the subindex 0 was actually part of this subset.